simple_rag/requirements.txt
chromadb
python-dotenv
google-genai
sentence-transformers
numpy<2